[233. Number of Digit One][Accepted]

+++ 233-Number-of-Digit-One/solution.java
@@ -2,7 +2,7 @@ public class Solution {
     public int countDigitOne(int n) {
         int ones = 0;
         for(long m = 1; m <= n; m *= 10) {
-            int a = n / m, b = n % m;
+            long a = n / m, b = n % m;
             ones += (a + 8) / 10 * m + ((a % 10 == 1) ? b + 1 : 0);
         }
